CAMS-85: cleared tech debt of code organization
