+++ user-interface/src/components/CaseList.tsx
@@ -1,6 +1,7 @@
 import { useState, useEffect } from 'react';
 import { useAppSelector } from '../store/store';
-import Api, { CaseListResponseData } from '../models/api';
+import Api from '../models/api';
+import { CaseListResponseData } from '../type-declarations/api';
 import './CaseList.scss';
 import MockApi from '../models/chapter11-mock.api.cases';
