+++ user-interface/src/type-declarations/basic.d.ts
@@ -1,0 +1,3 @@
+export type ObjectKeyVal = {
+  [key: string]: string | number;
+};
