+++ user-interface/src/type-declarations/chapter-15.d.ts
@@ -1,0 +1,13 @@
+import { ResponseData } from './api';
+
+export interface Chapter15Type {
+  caseNumber: string;
+  caseTitle: string;
+  dateFiled: string;
+}
+
+export interface Chapter15CaseListResponseData extends ResponseData {
+  body: {
+    caseList: Array<object>;
+  };
+}
